-Add AuthorizationController
